# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления заявками на системную нагрузку",
        "title": "System Load API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/system_load_strategies/add/{strategy_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет стратегию в заявку-черновик пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system_load_strategies"
                ],
                "summary": "Добавить стратегию в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка с добавленной стратегией",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.SystemLoadJSON"
                        }
                    },
                    "201": {
                        "description": "Создана новая заявка",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.SystemLoadJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Стратегия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Стратегия уже в заявке",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/system_load_strategies/{strategy_id}/{system_load_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет параметры стратегии в конкретной заявке (только черновик)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system_load_strategies"
                ],
                "summary": "Изменить данные стратегии в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "system_load_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.SystemLoadStrategyJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.SystemLoadStrategyJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь стратегии и заявки (только черновик)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system_load_strategies"
                ],
                "summary": "Удалить стратегию из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "system_load_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.SystemLoadJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "web_backend_internal_app_serializer.SystemLoadJSON": {
            "type": "object",
            "properties": {
                "completed_item_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "finish_date": {
                    "type": "string"
                },
                "forming_date": {
                    "type": "string"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "system_load_id": {
                    "type": "integer"
                }
            }
        },
        "web_backend_internal_app_serializer.SystemLoadStrategyJSON": {
            "type": "object",
            "properties": {
                "data_volume": {
                    "type": "integer"
                },
                "query_count": {
                    "type": "integer"
                },
                "response_time": {
                    "type": "number"
                },
                "strategy_id": {
                    "type": "integer"
                },
                "system_load_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}